add geofence "/getById" "/save"

+++ src/main/scala/cl/redd/objects/GeofenceOld.scala
@@ -2,14 +2,14 @@ package cl.redd.objects
 
 case class GeofenceOld (
 
-                         id		                   : Option[Int]     = None, //*
+                         id		                   : Option[Long]     = None, //*
                          name		                 : Option[String]  = None, //*
                          buffer	                 : Option[Int]     = None, //*
                          last_update_timestamp   : Option[Long]    = None,
                          alarm		               : Option[Boolean] = None, //*
                          latitude	               : Option[Double]  = None, //*
                          colour                  : Option[String]  = None,
-                         userid                  : Option[Int]     = None,
+                         userid                  : Option[Long]     = None,
                          longitude	             : Option[Double]  = None, //*
                          total		               : Option[Int]     = None, //*
                          typeId	                 : Option[Int]     = None, //*
